Simplify the `Monitor` to a one off `OrchestratorRegistration`
