+++ lib/nerves_hub/deployments/distributed/supervisor.ex
@@ -3,6 +3,8 @@ defmodule NervesHub.Deployments.Distributed.Supervisor do
 
   use Supervisor
 
+  alias NervesHub.Deployments.Distributed.OrchestratorRegistration
+
   def start_link(_) do
     Supervisor.start_link(__MODULE__, [], name: __MODULE__)
   end
@@ -10,7 +12,7 @@ defmodule NervesHub.Deployments.Distributed.Supervisor do
   def init(_) do
     children = [
       ProcessHub.child_spec(%ProcessHub{hub_id: :deployment_orchestrators}),
-      NervesHub.Deployments.Distributed.Monitor
+      OrchestratorRegistration
     ]
 
     Supervisor.init(children, strategy: :one_for_one)
